requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
python-pptx
